refactor(desktop): extract validate_project's checks into a validation module
`desktop.rs`'s `validate_project` method had grown to ~820 lines of inline
checks (disc, title, menu, titleset, build-settings) interleaved with their
helper functions and a 500-line test module (#48). Extracted into
`validation/{disc,title,menu,menu_action,menu_aspect,scene,titleset,
build_settings}.rs`, one file per rule category, orchestrated by
`validation::run()`.

`SpindleProject::validate_project` is now a one-line delegation to
`crate::validation::run(project)`. Check order, issue codes, and messages
are unchanged — this is a pure reorganisation, not a behavior change.
`validation` is gated `#[cfg(desktop)]` since the mobile build (`mobile.rs`)
delegates validation to the native mobile plugin instead.

Verified with `cargo build --workspace`, `cargo test --workspace` (189
passing — same tests, now under `validation::tests`), `cargo clippy
--workspace --all-targets -- -D warnings`, and `cargo fmt --check`.
